+++ examples/python/demo-pyapp-structure/pyapp/tests/test_lambda_example/__init__.py
@@ -1,0 +1,14 @@
+import unittest
+import importlib
+from pyapp.tests import ImportFromSourceContext
+
+with ImportFromSourceContext():
+    LAMBDA_HANDLER = importlib.import_module('lambdas.lambda_example.handler')
+
+
+class LambdaExampleLambdaTestCase(unittest.TestCase):
+    """Common setups for this lambda"""
+
+    def setUp(self) -> None:
+        self.HANDLER = LAMBDA_HANDLER.LambdaExample()
+